Move FileHashInvalidException to the files package.
